refactor: marketsApi endpoints
